Add global & scoped css reset
