+++ packages/strapi-parts/src/GlobalStyle/index.js
@@ -1,0 +1,1 @@
+export * from './GlobalStyle';
