+++ packages/strapi-parts/src/ScopedStyle/index.js
@@ -1,0 +1,1 @@
+export * from './ScopedStyle';
