+++ packages/strapi-parts/src/ThemeProvider/ThemeProvider.js
@@ -1,182 +1,26 @@
-import * as React from 'react';
 import PropTypes from 'prop-types';
-import { ThemeProvider as StyledThemeProvider, createGlobalStyle } from 'styled-components';
+import * as React from 'react';
+import { ThemeProvider as StyledThemeProvider } from 'styled-components';
+import { GlobalStyle } from '../GlobalStyle';
+import { ScopedStyle } from '../ScopedStyle';
 
-const GlobalStyle = createGlobalStyle`
- /* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-  html,
-  body,
-  div,
-  span,
-  applet,
-  object,
-  iframe,
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6,
-  p,
-  blockquote,
-  pre,
-  a,
-  abbr,
-  acronym,
-  address,
-  big,
-  cite,
-  code,
-  del,
-  dfn,
-  em,
-  img,
-  ins,
-  kbd,
-  q,
-  s,
-  samp,
-  small,
-  strike,
-  strong,
-  sub,
-  sup,
-  tt,
-  var,
-  b,
-  u,
-  i,
-  center,
-  dl,
-  dt,
-  dd,
-  ol,
-  ul,
-  li,
-  fieldset,
-  form,
-  label,
-  legend,
-  table,
-  caption,
-  tbody,
-  tfoot,
-  thead,
-  tr,
-  th,
-  td,
-  article,
-  aside,
-  canvas,
-  details,
-  embed,
-  figure,
-  figcaption,
-  footer,
-  header,
-  hgroup,
-  menu,
-  nav,
-  output,
-  ruby,
-  section,
-  summary,
-  time,
-  mark,
-  audio,
-  video {
-    margin: 0;
-    padding: 0;
-    border: 0;
-    font-size: 100%;
-    font: inherit;
-    vertical-align: baseline;
-  }
-  /* HTML5 display-role reset for older browsers */
-  article,
-  aside,
-  details,
-  figcaption,
-  figure,
-  footer,
-  header,
-  hgroup,
-  menu,
-  nav,
-  section {
-    display: block;
-  }
-  body {
-    line-height: 1;
-  }
-  ol,
-  ul {
-    list-style: none;
-  }
-  blockquote,
-  q {
-    quotes: none;
-  }
-  blockquote:before,
-  blockquote:after,
-  q:before,
-  q:after {
-    content: "";
-    content: none;
-  }
-  table {
-    border-collapse: collapse;
-    border-spacing: 0;
-  }
-  /* My styles */
-  *,
-  *:before,
-  *:after {
-    box-sizing: border-box;
-  }
-
-  *:focus-visible {
-    outline: 2px solid ${({ theme }) => theme.colors.primary600};
-    outline-offset: 2px;
-    box-shadow: revert;
-  }
-  /* Focusing the button with a mouse, touch, or stylus will show a subtle drop shadow. */
-  *:focus:not(:focus-visible) {
-    outline: none;
-  }
-
-  body,html{
-    height: 100%;
-  }
-  body {
-    font-family: --apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  }
-
-  button {
-    margin: 0;
-    padding: 0;
-    border: none;
-    background: none;
-    color: inherit;
-    font-size: 100%;
-    cursor: pointer;
-    font: inherit;
-  }
-`;
-
-export const ThemeProvider = ({ children, theme }) => {
+export const ThemeProvider = ({ children, theme, scopedCss = false }) => {
   return (
     <StyledThemeProvider theme={theme}>
-      <GlobalStyle />
-      {children}
+      {scopedCss ? (
+        <ScopedStyle>{children}</ScopedStyle>
+      ) : (
+        <>
+          <GlobalStyle />
+          {children}
+        </>
+      )}
     </StyledThemeProvider>
   );
 };
 
 ThemeProvider.propTypes = {
   children: PropTypes.node.isRequired,
+  scopedCss: PropTypes.bool,
   theme: PropTypes.object.isRequired,
 };
